🎨 style(index.css): remove display flex and place-items center from body The display flex and place-items center properties were removed from the body selector in the index.css file.
✨ feat(App.tsx): add AuthProvider component
The AuthProvider component was added to the App.tsx file to provide authentication context to the application.

🐛 fix(server.ts): change port variable case from lowercase port to uppercase PORT
The port variable was changed to uppercase PORT in the server.ts file to improve consistency with naming conventions.

✨ feat(server.ts): add support for process.env.PORT environment variable
The server.ts file was updated to add support for the process.env.PORT environment variable, which allows the application to run on any available port specified via the environment variable.

🎨 style(Drawer.tsx): add PapitoUI title and adjust styles
The Drawer.tsx file was updated to add a PapitoUI title and adjust the styles of the drawer.

🆕 feat(Post.tsx): add Post page component
🆕 feat(index.tsx): add Post page route
The Post page component is added to display a single post with the given id. The component uses the useQuery hook from react-query to fetch the post data from the server. The Post page route is added to the router to allow navigation to the Post page.