🎨 style(Drawer.tsx): adjust marginTop to 75vh The marginTop value was adjusted to 75vh to improve the visual appearance of the component.
🎨 style(WithStyleDrawer.tsx): change width to 0
The width of the drawer was changed to 0 to improve the visual appearance of the component.

🚀 feat(ThemeContext.tsx): add MuiThemeProvider component
The MuiThemeProvider component was added to provide a theme to the application.

🚀 feat(theme.ts): add default theme
A default theme was added to the application.

🚀 feat(MainLayout.tsx): add MuiThemeProvider component
The MuiThemeProvider component was added to provide a theme to the application.

🎨 style(Page.tsx): remove height and adjust paddingTop
The height property was removed and the paddingTop value was adjusted to improve the visual appearance of the component.

🚀 feat(main.tsx): add Fragment import